Fixed bug when creating PersonalInformation
When PersonalInformation.createFrom(request) was called, the method tried saving personal information before the address was properly set, therefore running into a SQL error. The createFrom() method was changed to accept an Address parameter. Corresponding code in services and their tests was also updated to reflect the method signature change.